+++ test/tensor_methods_test.rb
@@ -128,4 +128,8 @@ class TensorMethodsTest < Minitest::Test
     assert_equal [2, 3, 4], x.to_a
     assert_equal [1, 2, 3], y.to_a
   end
+
+  def test_cumsum_type
+    assert_equal :int64, Torch.tensor([1, 2, 3]).cumsum(dim: 0).dtype
+  end
 end